Rename phpunit fixture extension
