+++ phpunit.xml.dist
@@ -18,7 +18,7 @@
     </testsuites>
 
     <extensions>
-        <extension class="Cake\TestSuite\FixtureSchemaExtension" />
+        <extension class="Cake\TestSuite\Fixture\PHPUnitExtension"/>
     </extensions>
 
     <!-- Prevent coverage reports from looking in tests, vendors, config folders -->
